+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/invariants/group/GlobalConstraintDefinition.scala
@@ -14,7 +14,7 @@ import oscar.cbls.core._
   * @tparam T type of pre-computes used by the invariant
   * @tparam U type of the output of the invariant
   */
-abstract class GlobalConstraintDefinition[@specialized(Long) T : Manifest, @specialized(Long) U:Manifest](routes: ChangingSeqValue, v: Long)
+abstract class GlobalConstraintDefinition[@specialized(Long) T : Manifest, @specialized(Long) U:Manifest](routes: ChangingSeqValue, v: Int)
   extends Invariant with SeqNotificationTarget{
 
   val n = routes.maxValue+1L
